Json: strip content before parsing
I recently encountered an upstream server that includes several blank
lines in the response body before the JSON content and this causes
`JSON#parse` to return a `ParserError` for some reason. `#parse` can
handle leading/trailing blank lines, so I'm not sure why it fails
with this particular output. That said, calling `#strip` on the text
before it's parsed resolves this particular issue.